Server/back-end/cfehome/urls.py
from django.contrib import admin
from django.urls import path, include
from django.urls import path
from rest_framework.urls import urlpatterns
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('api.urls')),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/courses/', include('courses.urls')),
    path('api/quizzes/', include('quizzes.urls')),
    path('api/profiles/', include('profiles.urls')),
    path('api/', include('users.urls')),
]
