Server/requirements.txt
django==5.2.3
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pyyaml
requests
